+++ packages/ui/index.ts
@@ -27,6 +27,7 @@ export * from "./components/spinner";
 export * from "./components/table";
 export * from "./components/tabs";
 export * from "./components/textarea";
+export * from "./components/theme-provider";
 export * from "./components/toast";
 export * from "./components/tooltip";
 export * from "./lib";